refactor(park-keeper): 千行檔拆分收斂可維護性
- NavOverlay 1012 行破千行門檻：純搬移抽出 compassThemeStyles 設定與 useScreenWakeLock（926 行）
- Home 抽出 QuickCaptureCta 元件（658 到 592 行，低於 S0 後回漲基線）
- 三處均為零行為變更純搬移，介面與參數不動

測試：typecheck、vitest 586 passed、build、e2e 32 passed
