+++ apps/park-keeper/src/components/NavOverlay.tsx
@@ -23,7 +23,7 @@ import {
   Smartphone,
 } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
-import type { ThemeConfig, ThemeType, ParkingRecord } from '@app/park-keeper/types';
+import type { ThemeConfig, ParkingRecord } from '@app/park-keeper/types';
 import { useNavigation, getDirectionInfo } from '@app/park-keeper/hooks/useNavigation';
 import type { DirectionIconType } from '@app/park-keeper/hooks/useNavigation';
 import {
@@ -50,100 +50,14 @@ import {
   ON_PRIMARY_COLOR,
 } from '@app/park-keeper/config/colors';
 import { useModalDialog } from '@app/park-keeper/hooks/useModalDialog';
+import { useScreenWakeLock } from '@app/park-keeper/hooks/useScreenWakeLock';
+import { COMPASS_THEME_STYLES } from '@app/park-keeper/config/compassThemeStyles';
 import PhotoViewerModal from './PhotoViewerModal';
 
 const MiniMap = lazy(() => import('./MiniMap'));
 
-// ---------------------------------------------------------------------------
-// 主題差異化樣式參數（brief：Nitro 霓虹描邊／Kawaii 粉彩圓潤／Zen 細線極簡／Classic 銅色調）
-// ---------------------------------------------------------------------------
-interface CompassThemeStyle {
-  /** 刻度線寬係數。 */
-  tickWidthScale: number;
-  /** 刻度端點形狀（Kawaii 圓頭）。 */
-  tickLinecap: 'round' | 'butt';
-  /** 外環線寬。 */
-  outerRingWidth: number;
-  /** 外環不透明度。 */
-  outerRingOpacity: number;
-  /** 盤面霓虹光暈（SVG drop-shadow filter 強度，0 = 無）。 */
-  neonGlowRadius: number;
-  /** 楔形未對準時不透明度。 */
-  wedgeIdleOpacity: number;
-}
-
-const COMPASS_THEME_STYLES: Record<ThemeType, CompassThemeStyle> = {
-  racing: {
-    tickWidthScale: 1,
-    tickLinecap: 'butt',
-    outerRingWidth: 1.6,
-    outerRingOpacity: 0.55,
-    neonGlowRadius: 6,
-    wedgeIdleOpacity: 0.4,
-  },
-  cute: {
-    tickWidthScale: 1.5,
-    tickLinecap: 'round',
-    outerRingWidth: 2.5,
-    outerRingOpacity: 0.25,
-    neonGlowRadius: 0,
-    // 粉彩 primary 飽和度低，楔形 idle 提高補償可見度。
-    wedgeIdleOpacity: 0.65,
-  },
-  minimalist: {
-    tickWidthScale: 0.8,
-    tickLinecap: 'butt',
-    outerRingWidth: 0.8,
-    outerRingOpacity: 0.14,
-    neonGlowRadius: 0,
-    wedgeIdleOpacity: 0.32,
-  },
-  literary: {
-    tickWidthScale: 1.1,
-    tickLinecap: 'butt',
-    outerRingWidth: 2,
-    outerRingOpacity: 0.35,
-    neonGlowRadius: 0,
-    wedgeIdleOpacity: 0.38,
-  },
-};
-
 /** 對準觸覺脈衝（Android；iOS 不支援 vibrate 靜默降級）。 */
 const ALIGNED_VIBRATE_PATTERN = [30, 40, 30];
-
-// ---------------------------------------------------------------------------
-// Screen Wake Lock：導航時防熄屏；hidden 釋放、回前景重新取得（iOS 18.4+ 修復）。
-// ---------------------------------------------------------------------------
-function useScreenWakeLock() {
-  useEffect(() => {
-    if (typeof navigator === 'undefined' || !('wakeLock' in navigator)) return undefined;
-
-    let sentinel: WakeLockSentinel | null = null;
-    let disposed = false;
-
-    const request = async () => {
-      try {
-        sentinel = await navigator.wakeLock.request('screen');
-      } catch {
-        // 低電量模式或平台拒絕：靜默降級，不影響導航功能。
-        sentinel = null;
-      }
-    };
-
-    const handleVisibility = () => {
-      if (document.visibilityState === 'visible' && !disposed) void request();
-    };
-
-    void request();
-    document.addEventListener('visibilitychange', handleVisibility);
-
-    return () => {
-      disposed = true;
-      document.removeEventListener('visibilitychange', handleVisibility);
-      void sentinel?.release().catch(() => undefined);
-    };
-  }, []);
-}
 
 // ---------------------------------------------------------------------------
 // DirectionIcon – Lucide icon mapped from DirectionIconType
